ingestion_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
influxdb-client==1.38.0
psutil==5.9.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
